pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "dask_azure_storage_blob"
authors = [{name = "Tom Augspurger", email = "taugspurger@microsoft.com"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "azure-storage-blob >=12.8.1",
    "dask >=2021.4.0",
]